+++ code/crates/engine/src/host.rs
@@ -8,7 +8,6 @@ use malachitebft_core_consensus::{Role, VoteExtensionError};
 use malachitebft_core_types::{CommitCertificate, Context, Round, ValueId, VoteExtensions};
 use malachitebft_sync::{PeerId, RawDecidedValue};
 
-use crate::consensus::ConsensusRef;
 use crate::util::streaming::StreamMessage;
 
 pub use malachitebft_core_consensus::{LocallyProposedValue, ProposedValue};
@@ -16,25 +15,57 @@ pub use malachitebft_core_consensus::{LocallyProposedValue, ProposedValue};
 /// A reference to the host actor.
 pub type HostRef<Ctx> = ActorRef<HostMsg<Ctx>>;
 
+/// What to do next after a decision.
+#[derive_where(Debug)]
+pub enum Next<Ctx: Context> {
+    /// Start at the given height with the given validator set.
+    Start(Ctx::Height, Ctx::ValidatorSet),
+
+    /// Restart at the given height with the given validator set.
+    Restart(Ctx::Height, Ctx::ValidatorSet),
+}
+
 /// Messages that need to be handled by the host actor.
 #[derive_where(Debug)]
 pub enum HostMsg<Ctx: Context> {
-    /// Consensus is ready
-    ConsensusReady(ConsensusRef<Ctx>),
+    /// Notifies the application that consensus is ready.
+    ///
+    /// The application MUST reply with a message to instruct
+    /// consensus to start at a given height.
+    ConsensusReady {
+        /// Use this reply port to instruct consensus to start the first height.
+        reply_to: RpcReplyPort<(Ctx::Height, Ctx::ValidatorSet)>,
+    },
 
     /// Consensus has started a new round.
     StartedRound {
+        /// The height at which the round started.
         height: Ctx::Height,
+        /// The round number that started.
         round: Round,
+        /// The address of the proposer for this round.
         proposer: Ctx::Address,
+        /// The role of the node in this round.
         role: Role,
+        /// Use this reply port to send the undecided values that were already seen for this
+        /// round. This is needed when recovering from a crash.
+        ///
+        /// The application MUST reply immediately with the values it has, or with an empty vector.
+        reply_to: RpcReplyPort<Vec<ProposedValue<Ctx>>>,
     },
 
     /// Request to build a local value to propose
+    ///
+    /// The application MUST reply to this message with the requested value
+    /// within the specified timeout duration.
     GetValue {
+        /// The height at which the value should be proposed.
         height: Ctx::Height,
+        /// The round in which the value should be proposed.
         round: Round,
+        /// The amount of time the application has to build the value.
         timeout: Duration,
+        /// Use this reply port to send the value that was built.
         reply_to: RpcReplyPort<LocallyProposedValue<Ctx>>,
     },
 
@@ -45,9 +76,13 @@ pub enum HostMsg<Ctx: Context> {
     /// This data is opaque to the consensus algorithm but can contain application-specific information.
     /// The proposer of the next block will receive all vote extensions along with the commit certificate.
     ExtendVote {
+        /// The height at which the vote is being extended.
         height: Ctx::Height,
+        /// The round in which the vote is being extended.
         round: Round,
+        /// The ID of the value that is being voted on.
         value_id: ValueId<Ctx>,
+        /// The vote extension to be added to the vote, if any.
         reply_to: RpcReplyPort<Option<Ctx::Extension>>,
     },
 
@@ -56,39 +91,70 @@ pub enum HostMsg<Ctx: Context> {
     /// If the vote extension is deemed invalid, the vote it was part of
     /// will be discarded altogether.
     VerifyVoteExtension {
+        /// The height for which the vote is.
         height: Ctx::Height,
+        /// The round for which the vote is.
         round: Round,
+        /// The ID of the value that the vote extension is for.
         value_id: ValueId<Ctx>,
+        /// The vote extension to verify.
         extension: Ctx::Extension,
+        /// Use this reply port to send the result of the verification.
         reply_to: RpcReplyPort<Result<(), VoteExtensionError>>,
     },
 
-    /// Request to restream an existing block/value from Driver
+    /// Requests the application to re-stream a proposal that it has already seen.
+    ///
+    /// The application MUST re-publish again all the proposal parts pertaining
+    /// to that value by sending [`NetworkMsg::PublishProposalPart`] messages through
+    /// the [`Channels::network`] channel.
     RestreamValue {
+        /// The height at which the value was proposed.
         height: Ctx::Height,
+        /// The round in which the value was proposed.
         round: Round,
+        /// The round in which the value was valid.
         valid_round: Round,
+        /// The address of the proposer of the value.
         address: Ctx::Address,
+        /// The ID of the value to restream.
         value_id: ValueId<Ctx>,
     },
 
-    /// Request the earliest block height in the block store
+    /// Requests the earliest height available in the history maintained by the application.
+    ///
+    /// The application MUST respond with its earliest available height.
     GetHistoryMinHeight { reply_to: RpcReplyPort<Ctx::Height> },
 
-    /// ProposalPart received <-- consensus <-- gossip
+    /// Notifies the application that consensus has received a proposal part over the network.
+    ///
+    /// If this part completes the full proposal, the application MUST respond
+    /// with the complete proposed value. Otherwise, it MUST respond with `None`.
     ReceivedProposalPart {
         from: PeerId,
         part: StreamMessage<Ctx::ProposalPart>,
         reply_to: RpcReplyPort<ProposedValue<Ctx>>,
     },
 
-    /// Get the validator set at a given height
+    /// Requests the validator set for a specific height
     GetValidatorSet {
         height: Ctx::Height,
         reply_to: RpcReplyPort<Option<Ctx::ValidatorSet>>,
     },
 
-    /// Consensus has decided on a value.
+    /// Notifies the application that consensus has decided on a value.
+    ///
+    /// This message includes a commit certificate containing the ID of
+    /// the value that was decided on, the height and round at which it was decided,
+    /// and the aggregated signatures of the validators that committed to it.
+    /// It also includes to the vote extensions received for that height.
+    ///
+    /// In response to this message, the application MUST send a [`Next`]
+    /// message back to consensus, instructing it to either start the next height if
+    /// the application was able to commit the decided value, or to restart the current height
+    /// otherwise.
+    ///
+    /// If the application does not reply, consensus will stall.
     Decided {
         /// The commit certificate containing the ID of the value that was decided on,
         /// the the height and round at which it was decided, and the aggregated signatures
@@ -98,24 +164,36 @@ pub enum HostMsg<Ctx: Context> {
         /// Vote extensions that were received for this height.
         extensions: VoteExtensions<Ctx>,
 
-        /// Reference to the `Consensus` actor for starting a new height.
-        consensus: ConsensusRef<Ctx>,
+        /// Use this reply port to instruct consensus to start the next height.
+        reply_to: RpcReplyPort<Next<Ctx>>,
     },
 
-    // Retrieve decided value from the block store
+    /// Requests a previously decided value from the application's storage.
+    ///
+    /// The application MUST respond with that value if available, or `None` otherwise.
     GetDecidedValue {
+        /// Height of the decided value to retrieve
         height: Ctx::Height,
+        /// Channel for sending back the decided value
         reply_to: RpcReplyPort<Option<RawDecidedValue<Ctx>>>,
     },
 
-    // Process a value synced from another node via the ValueSync protocol.
-    //
-    // If the encoded value within is valid, the host MUST reply with that value to be proposed.
+    /// Notifies the application that a value has been synced from the network.
+    /// This may happen when the node is catching up with the network.
+    ///
+    /// If a value can be decoded from the bytes provided, then the application MUST reply
+    /// to this message with the decoded value. Otherwise, it MUST reply with `None`.
     ProcessSyncedValue {
+        /// Height of the synced value
         height: Ctx::Height,
+        /// Round of the synced value
         round: Round,
-        validator_address: Ctx::Address,
+        /// Address of the original proposer
+        proposer: Ctx::Address,
+        /// Raw encoded value data
         value_bytes: Bytes,
+        /// Channel for sending back the proposed value, if successfully decoded
+        /// or `None` if the value could not be decoded
         reply_to: RpcReplyPort<ProposedValue<Ctx>>,
     },
 }