+++ code/examples/channel/src/app.rs
@@ -1,6 +1,7 @@
 use std::time::Duration;
 
 use eyre::eyre;
+use malachitebft_app_channel::app::engine::host::Next;
 use tokio::time::sleep;
 use tracing::{error, info};
 
@@ -8,7 +9,7 @@ use malachitebft_app_channel::app::streaming::StreamContent;
 use malachitebft_app_channel::app::types::core::{Height as _, Round, Validity};
 use malachitebft_app_channel::app::types::sync::RawDecidedValue;
 use malachitebft_app_channel::app::types::{LocallyProposedValue, ProposedValue};
-use malachitebft_app_channel::{AppMsg, Channels, ConsensusMsg, NetworkMsg};
+use malachitebft_app_channel::{AppMsg, Channels, NetworkMsg};
 use malachitebft_test::{Height, TestContext};
 
 use crate::state::{decode_value, encode_value, State};
@@ -194,9 +195,12 @@ pub async fn run(state: &mut State, channels: &mut Channels<TestContext>) -> eyr
                 // When that happens, we store the decided value in our store
                 match state.commit(certificate, extensions).await {
                     Ok(_) => {
+                        // Sleep a bit to slow down the app.
+                        sleep(Duration::from_millis(500)).await;
+
                         // And then we instruct consensus to start the next height
                         if reply
-                            .send(ConsensusMsg::StartHeight(
+                            .send(Next::Start(
                                 state.current_height,
                                 state.get_validator_set(state.current_height).clone(),
                             ))
@@ -206,12 +210,15 @@ pub async fn run(state: &mut State, channels: &mut Channels<TestContext>) -> eyr
                         }
                     }
                     Err(_) => {
+                        let height = state.current_height;
+
                         // Commit failed, restart the height
-                        error!("Commit failed, restarting height {}", state.current_height);
+                        error!("Commit failed, restarting height {height}");
+
                         if reply
-                            .send(ConsensusMsg::RestartHeight(
-                                state.current_height,
-                                state.get_validator_set(state.current_height).clone(),
+                            .send(Next::Restart(
+                                height,
+                                state.get_validator_set(height).clone(),
                             ))
                             .is_err()
                         {
@@ -219,7 +226,6 @@ pub async fn run(state: &mut State, channels: &mut Channels<TestContext>) -> eyr
                         }
                     }
                 }
-                sleep(Duration::from_millis(500)).await;
             }
 
             // It may happen that our node is lagging behind its peers. In that case,